+++ core/module/theme/theme.php
@@ -490,6 +490,7 @@ class theme extends common {
 		if($this->isPost()) {
 			$this->setData(['theme', 'menu', [
 				'backgroundColor' => $this->getInput('themeMenuBackgroundColor'),
+				'backgroundColorSub' => $this->getInput('themeMenuBackgroundColorSub'),
 				'font' => $this->getInput('themeMenuFont'),
 				'fontSize' => $this->getInput('themeMenuFontSize'),
 				'fontWeight' => $this->getInput('themeMenuFontWeight'),